+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ## Refactoring
 - Improve mayEditStore() to fail faster !1311 @jofranz
+- Moved the button for new stores to vue store list !1282 @chriswalg
 
 ## Bugfixes
 - Improved the banner on welcome page for mobile devices !1329 @chriswalg